chil/tools/services/empirical_service.py
"""
Empirical Framework Tools Service Implementation.
"""

import grpc
import time
import asyncio
from concurrent import futures
from typing import Dict, List, Any

import z3
from duckdb import connect as duckdb_connect
from oxigraph import Store

from ..protos import tools_pb2
from ..protos import tools_pb2_grpc


class EmpiricalToolsServicer(tools_pb2_grpc.EmpiricalToolsServicer):
    """gRPC service for empirical framework tools."""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.duckdb_conn = duckdb_connect(config.get('duckdb_path', ':memory:'))
        self.rdf_store = Store(config.get('oxigraph_path', ':memory:'))
        self._init_databases()
    
    def _init_databases(self):
        """Initialize databases with schemas."""
        # DuckDB fact tables
        self.duckdb_conn.execute("""
            CREATE TABLE IF NOT EXISTS facts (
                id INTEGER PRIMARY KEY,
                claim TEXT,
                source TEXT,
                confidence DOUBLE,
                timestamp TIMESTAMP,
                evidence JSON
            )
        """)
        
        # Create indexes
        self.duckdb_conn.execute("""
            CREATE INDEX IF NOT EXISTS idx_facts_claim 
            ON facts(claim)
        """)
    
    def CheckLogicalConsistency(self, request, context):
        """Check logical consistency using Z3."""
        start_time = time.time()
        
        try:
            solver = z3.Solver()
            solver.set("timeout", request.timeout_ms)
            
            # Parse formula
            formula = z3.parse_smt2_string(request.formula)
            solver.add(formula)
            
            # Add constraints
            for constraint in request.constraints:
                solver.add(z3.parse_smt2_string(constraint))
            
            # Check satisfiability
            result = solver.check()
            
            response = tools_pb2.Z3Response()
            response.satisfiable = (result == z3.sat)
            
            if response.satisfiable:
                model = solver.model()
                for decl in model.decls():
                    response.model[str(decl)] = str(model[decl])
            
            if result == z3.unsat:
                response.proof = str(solver.proof())
            
            return response
            
        except Exception as e:
            context.abort(grpc.StatusCode.INTERNAL, f"Z3 error: {str(e)}")
    
    def ProveTheorem(self, request, context):
        """Prove theorem using Lean (simplified interface)."""
        # This is a placeholder - real Lean integration would use lean4 Python bindings
        response = tools_pb2.LeanResponse()
        
        # Simulate theorem proving
        if "trivial" in request.tactic:
            response.proven = True
            response.proof = f"Proof of {request.theorem} by {request.tactic}"
            response.steps.extend(["Apply axioms", "Use tactic", "QED"])
        else:
            response.proven = False
            response.proof = "Unable to prove"
        
        return response
    
    def QueryFactDatabase(self, request, context):
        """Query fact database using DuckDB."""
        try:
            # Execute query with parameters
            if request.parameters:
                result = self.duckdb_conn.execute(
                    request.query, 
                    [request.parameters[k] for k in sorted(request.parameters.keys())]
                ).fetchall()
            else:
                result = self.duckdb_conn.execute(request.query).fetchall()
            
            response = tools_pb2.DuckDBResponse()
            response.row_count = len(result)
            
            # Get column names
            columns = [desc[0] for desc in self.duckdb_conn.description]
            
            # Convert results
            for row in result:
                row_msg = tools_pb2.DuckDBResponse.Row()
                for i, col in enumerate(columns):
                    row_msg.columns[col] = str(row[i])
                response.results.append(row_msg)
            
            return response
            
        except Exception as e:
            context.abort(grpc.StatusCode.INTERNAL, f"DuckDB error: {str(e)}")
    
    def QueryKnowledgeGraph(self, request, context):
        """Query knowledge graph using SPARQL."""
        try:
            # Execute SPARQL query
            results = self.rdf_store.query(request.query)
            
            response = tools_pb2.SparqlResponse()
            
            for result in results:
                binding = tools_pb2.SparqlResponse.Binding()
                for var, value in result.items():
                    binding.variables[var] = str(value)
                response.bindings.append(binding)
            
            return response
            
        except Exception as e:
            context.abort(grpc.StatusCode.INTERNAL, f"SPARQL error: {str(e)}")


def serve(port: int = 50051, config: Dict[str, Any] = None):
    """Start the empirical tools gRPC server."""
    config = config or {}
    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
    tools_pb2_grpc.add_EmpiricalToolsServicer_to_server(
        EmpiricalToolsServicer(config), server
    )
    server.add_insecure_port(f'[::]:{port}')
    server.start()
    return server


if __name__ == '__main__':
    server = serve()
    server.wait_for_termination()