chil/tools/requirements.txt
# gRPC and Protocol Buffers
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-health-checking>=1.60.0

# Empirical Framework Tools
z3-solver>=4.12.0
duckdb>=0.9.0
oxigraph>=0.3.0

# Contextual Framework Tools  
spacy>=3.7.0
gensim>=4.3.0
faiss-cpu>=1.7.4
textblob>=0.17.0

# Consistency Framework Tools
# SQLite comes with Python
souffle>=2.4  # Install separately: https://souffle-lang.github.io/install

# Power Dynamics Framework Tools
networkx>=3.2
scikit-learn>=1.3.0

# Utility Framework Tools
xgboost>=2.0.0
lightgbm>=4.1.0
ortools>=9.7.0
numba>=0.58.0

# Evolution Framework Tools
deap>=1.4.0
jax>=0.4.20
jaxlib>=0.4.20
ray[rllib]>=2.8.0

# Monitoring and Configuration
pyyaml>=6.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# Performance
numpy>=1.24.0
pandas>=2.1.0
pyarrow>=14.0.0
lmdb>=1.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0